# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Restaurant Suite API
  version: 0.1.0
  description: REST API versioned under /api/v1.
paths:
...
  /api/v1/tables/{id}:
    put:
      operationId: updateTable
      x-required-permission: floor.layout.write
      security:
        - staffSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
        - name: If-Match
          in: header
          required: true
          description: Current table version
          schema:
            type: string
            pattern: ^"?[1-9][0-9]*"?$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTableRequest"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "428":
          $ref: "#/components/responses/Error"
...
  /api/v1/locations/{locationId}/orders/{orderId}/lines:
    post:
      operationId: addOrderLines
      x-required-permission: orders.lines.add
      security:
        - staffSession: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
        - name: orderId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderLinesRequest"
      responses:
        "201":
          description: Server-priced lines
          content:
            application/json:
              schema:
                type: object
        "409":
          $ref: "#/components/responses/Error"
...
  /api/v1/organizations/{org_id}/customers:
    post:
      summary: Register a new customer
      tags:
        - Customers
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /api/v1/organizations/{org_id}/customer-sessions:
    post:
      summary: Login customer
      tags:
        - Customers
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCustomerRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSessionResponse"
        "401":
          description: Invalid credentials
  /api/v1/organizations/{org_id}/customer-sessions/current:
    get:
      summary: Get current customer profile
      tags:
        - Customers
      security:
        - CustomerBearerAuth: []
      parameters:
        - name: org_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "401":
          description: Unauthorized
  /api/v1/locations/{loc_id}/online-orders/checkout:
    post:
      summary: Checkout online order
      tags:
        - OnlineOrdering
      security:
        - {}
        - CustomerBearerAuth: []
      parameters:
        - name: loc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlineCheckoutRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnlineCheckoutResponse"
  /api/v1/locations/{loc_id}/online-orders/{order_id}:
    get:
      summary: Get online order detail
      tags:
        - OnlineOrdering
      security:
        - {}
        - CustomerBearerAuth: []
      parameters:
        - name: loc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: order_token
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnlineOrderDetailResponse"
  /api/v1/locations/{loc_id}/orders/{order_id}/fulfillment/dispatch:
    post:
      summary: Dispatch delivery order
      tags:
        - OnlineOrdering
      security:
        - StaffBearerAuth: []
      parameters:
        - name: loc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FulfillmentActionRequest"
      responses:
        "200":
          description: Dispatched
        "409":
          description: Conflict
...
components:
  securitySchemes:
    staffSession:
      type: http
      scheme: bearer
      bearerFormat: opaque
    guestSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-guest
    CustomerBearerAuth:
      type: http
      scheme: bearer
      description: Customer session token (e.g. `customer.<org_id>.<secret>`)
...
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    Uuid:
      type: string
      format: uuid
...
    Table:
      type: object
      additionalProperties: false
      required:
        - id
        - location_id
        - area_id
        - name
        - min_capacity
        - max_capacity
        - pos_x
        - pos_y
        - status
        - version
        - created_at
        - updated_at
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        location_id:
          $ref: "#/components/schemas/Uuid"
        area_id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        min_capacity:
          type: integer
        max_capacity:
          type: integer
        pos_x:
          type: integer
        pos_y:
          type: integer
        status:
          type: string
          enum:
            - AVAILABLE
            - OCCUPIED
            - NEEDS_CLEANING
            - OUT_OF_ORDER
        version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateTableRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        area_id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
          minLength: 1
          maxLength: 160
        min_capacity:
          type: integer
          minimum: 1
        max_capacity:
          type: integer
          minimum: 1
        pos_x:
          type: integer
        pos_y:
          type: integer
...
    Order:
      type: object
      additionalProperties: true
      required:
        - id
        - location_id
        - visit_id
        - order_type
        - status
        - version
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        location_id:
          $ref: "#/components/schemas/Uuid"
        visit_id:
          $ref: "#/components/schemas/Uuid"
        order_type:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - HELD
            - SENT
            - COMPLETED
            - CANCELLED
        version:
          type: integer
    OrderLine:
      type: object
      additionalProperties: true
      required:
        - id
        - order_id
        - account_id
        - product_id
        - quantity
        - unit_price
        - status
        - version
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        order_id:
          $ref: "#/components/schemas/Uuid"
        account_id:
          $ref: "#/components/schemas/Uuid"
        product_id:
          $ref: "#/components/schemas/Uuid"
        quantity:
          type: integer
        unit_price:
          type: integer
          description: Server-computed cents
        status:
          type: string
          enum:
            - DRAFT
            - HELD
            - SENT
            - PREPARING
            - READY
            - FULFILLED
            - CANCELLED
            - VOIDED
        version:
          type: integer
...
    AddOrderLinesRequest:
      type: object
      additionalProperties: false
      required:
        - lines
      properties:
        lines:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required:
              - account_id
              - product_id
              - quantity
            properties:
              account_id:
                $ref: "#/components/schemas/Uuid"
              product_id:
                $ref: "#/components/schemas/Uuid"
              variant_id:
                $ref: "#/components/schemas/Uuid"
              modifier_ids:
                type: array
                items:
                  $ref: "#/components/schemas/Uuid"
              seat_number:
                type:
                  - integer
                  - "null"
              course_name:
                type:
                  - string
                  - "null"
              quantity:
                type: integer
                minimum: 1
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: true
          required:
            - status
            - code
            - message
            - request_id
          properties:
            status:
              type: integer
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details: {}
...
    Customer:
      type: object
      required:
        - id
        - organization_id
        - email
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CustomerSessionResponse:
      type: object
      required:
        - token
        - customer
      properties:
        token:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
    RegisterCustomerRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        phone:
          type: string
    LoginCustomerRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    OnlineCheckoutRequest:
      type: object
      required:
        - items
        - fulfillment_type
        - customer_name
        - customer_email
        - customer_phone
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutItem"
        fulfillment_type:
          type: string
          enum:
            - PICKUP
            - DELIVERY
        scheduled_for:
          type: string
          format: date-time
          nullable: true
        customer_name:
          type: string
        customer_email:
          type: string
          format: email
        customer_phone:
          type: string
        delivery_address:
          type: object
          additionalProperties: true
          nullable: true
    OnlineCheckoutResponse:
      type: object
      required:
        - order_id
      properties:
        order_id:
          type: string
          format: uuid
        order_token:
          type: string
          nullable: true
    OrderFulfillment:
      type: object
      required:
        - status
        - fulfillment_type
      properties:
        status:
          type: string
          enum:
            - PENDING
            - OUT_FOR_DELIVERY
            - DELIVERED
            - CANCELLED
        fulfillment_type:
          type: string
          enum:
            - PICKUP
            - DELIVERY
    OnlineOrderDetailResponse:
      type: object
      required:
        - order
        - lines
        - fulfillment
      properties:
        order:
          $ref: "#/components/schemas/Order"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
        fulfillment:
          $ref: "#/components/schemas/OrderFulfillment"
    FulfillmentActionRequest:
      type: object
      required:
        - version
      properties:
        version:
          type: integer
    CheckoutModifier:
      type: object
      required:
        - modifier_id
        - quantity
      properties:
        modifier_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    CheckoutItem:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
        notes:
          type: string
          nullable: true
        modifiers:
          type: array
          items:
            $ref: "#/components/schemas/CheckoutModifier"